Fixed bug in HMM similarity code.
There was an error in the compute_hmm_similarity function. The indexing
was not correct. The indices for the HMMs in the tiedlist and the
indices in the HMM set of the model are not the same. I needed to be
careful mapping between the two indices and using each when appropriate.