Added Keyword scripts and fixed output.
For modify_kw_hitlist.py, there was a bug in the writing of the output
to a file. It only worked to stdout.

Added to new keyword scripts.
subset_keywords.py: Takes a keyword definition and creates a new
definition containing only a subset of those keywords. Useful for making
the IV/OOV subsets.
create_keyword_features.py: Creates a set of keyword specific features.