Added create_training_features.py
Created the create_training_features.py script. Given a set of hitlists,
it merges them together to create a set of features that can be used to
train an ML algorithm.

Modified the kwhit and hitlist classes. The merging code is now inside
the classes instead of in a separate script.